+++ src/main/java/seedu/address/logic/parser/SearchCommandParser.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 import seedu.address.logic.commands.SearchCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.TagContainsKeywordsPredicate;
+import seedu.address.model.types.common.TagContainsKeywordsPredicate;
 
 /**
  * Parses input arguments and creates a new SearchCommand object
