--- src/main/java/seedu/address/model/event/Event.java
@@ -1,103 +1,0 @@
-package seedu.address.model.event;
-
-import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
-
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-
-import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Name;
-import seedu.address.model.tag.Tag;
-
-/**
- * Represents a Person in the address book.
- * Guarantees: details are present and not null, field values are validated, immutable.
- */
-public class Event {
-
-    // Identity fields
-    private final Name name;
-
-    // Data fields
-    private final Address address;
-    private final Set<Tag> tags = new HashSet<>();
-
-    /**
-     * Every field must be present and not null.
-     */
-    public Event(Name name, Address address, Set<Tag> tags) {
-        requireAllNonNull(name, address, tags);
-        this.name = name;
-        this.address = address;
-        this.tags.addAll(tags);
-    }
-
-    public Name getName() {
-        return name;
-    }
-
-    public Address getAddress() {
-        return address;
-    }
-
-    /**
-     * Returns an immutable tag set, which throws {@code UnsupportedOperationException}
-     * if modification is attempted.
-     */
-    public Set<Tag> getTags() {
-        return Collections.unmodifiableSet(tags);
-    }
-
-    /**
-     * Returns true if both persons have the same name.
-     * This defines a weaker notion of equality between two persons.
-     */
-    public boolean isSamePerson(Event otherEvent) {
-        if (otherEvent == this) {
-            return true;
-        }
-
-        return otherEvent != null
-                && otherEvent.getName().equals(getName());
-    }
-
-    /**
-     * Returns true if both persons have the same identity and data fields.
-     * This defines a stronger notion of equality between two persons.
-     */
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof Event)) {
-            return false;
-        }
-
-        Event otherEvent = (Event) other;
-        return name.equals(otherEvent.name)
-                && address.equals(otherEvent.address)
-                && tags.equals(otherEvent.tags);
-    }
-
-    @Override
-    public int hashCode() {
-        // use this method for custom fields hashing instead of implementing your own
-        return Objects.hash(name, address, tags);
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .add("name", name)
-                .add("address", address)
-                .add("tags", tags)
-                .toString();
-    }
-
-}
